test(tools): besu multi-party AIO container image
Works with a modified (forked) version of quorum-dev-quickstart
See https://github.com/ConsenSys/quorum-dev-quickstart/pull/72

On the fork we added an extra commit that removes the usage of
tput (which is a low impact change because all
it is being used for is to make text in the terminal bold for cosmetic
purposes).

With this change the container now works fine. The next issue up is
that it uses 15 GB RAM idling and 20% CPU (on a 6 core machine..)

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>