test(connector-besu): private contract tx visibility check
Note that this checks the visibility of the transactions
not the contract itself.
Everyone can invoke the contract that was deployed privately,
but the transactions executed by the contract can only be
seen by the given privacy group that has access.

The privacy groups members are determined by the union set
of the privateFor and privateFrom parameters.

The two tests verify the above, one with vanilla web3-eea
and another one with the Cactus connector.

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>